+++ builder/amazon/examples/nginx_server/packages.sh
@@ -1,4 +1,0 @@
-echo "installing nginx "
-sudo apt-get update
-sudo apt-get install nginx -y
-sudo service nginx restart