+++ builder/proxmox/common/builder.go
@@ -2,7 +2,6 @@ package proxmox
 
 import (
 	"context"
-	"crypto/tls"
 	"errors"
 	"fmt"
 
@@ -35,15 +34,7 @@ type Builder struct {
 
 func (b *Builder) Run(ctx context.Context, ui packersdk.Ui, hook packersdk.Hook, state multistep.StateBag) (packersdk.Artifact, error) {
 	var err error
-	tlsConfig := &tls.Config{
-		InsecureSkipVerify: b.config.SkipCertValidation,
-	}
-	b.proxmoxClient, err = proxmox.NewClient(b.config.proxmoxURL.String(), nil, tlsConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	err = b.proxmoxClient.Login(b.config.Username, b.config.Password, "")
+	b.proxmoxClient, err = newProxmoxClient(b.config)
 	if err != nil {
 		return nil, err
 	}
@@ -61,12 +52,7 @@ func (b *Builder) Run(ctx context.Context, ui packersdk.Ui, hook packersdk.Hook,
 		&stepStartVM{
 			vmCreator: b.vmCreator,
 		},
-		&commonsteps.StepHTTPServer{
-			HTTPDir:     b.config.HTTPDir,
-			HTTPPortMin: b.config.HTTPPortMin,
-			HTTPPortMax: b.config.HTTPPortMax,
-			HTTPAddress: b.config.HTTPAddress,
-		},
+		commonsteps.HTTPServerFromHTTPConfig(&b.config.HTTPConfig),
 		&stepTypeBootCommand{
 			BootConfig: b.config.BootConfig,
 			Ctx:        b.config.Ctx,
@@ -170,6 +156,9 @@ func getVMIP(state multistep.StateBag) (string, error) {
 	for _, iface := range ifs {
 		for _, addr := range iface.IPAddresses {
 			if addr.IsLoopback() {
+				continue
+			}
+			if addr.To4() == nil {
 				continue
 			}
 			return addr.String(), nil